requirements.txt
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
fonttools==4.58.4
fsspec==2024.6.1
imageio==2.37.0
Jinja2==3.1.4
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pillow==11.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
sympy==1.13.3
tifffile==2025.5.10
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.67.1
typing_extensions==4.12.2
